rename Test to TestRunDTO and put it in shared-models instead of applicant-server
