+++ applicant-server/src/main/kotlin/ut/isep/interview/code_execution/CodeExecutor.kt
@@ -1,6 +1,6 @@
 package ut.isep.interview.code_execution
 
-import ut.isep.interview.code_execution.dto.Test
+import dto.execution.TestRunDTO
 import ut.isep.interview.code_execution.utils.TestResult
 import java.io.File
 
@@ -8,5 +8,5 @@ interface CodeExecutor {
 
     fun startContainer(inviteId: String, container: File)
 
-    fun runTest(inviteId: String, test: Test): List<TestResult>
+    fun runTest(inviteId: String, test: TestRunDTO): List<TestResult>
 }
