+++ shared-models/src/main/kotlin/dto/execution/TestRunDTO.kt
@@ -1,6 +1,6 @@
-package ut.isep.interview.code_execution.dto
+package dto.execution
 
-data class Test(
+data class TestRunDTO(
     val code: String,
     val codeFileName: String?,
     val test: String?,
